Wp8 downloadresume

+++ windows-client/FileTransfer.cs
@@ -20,6 +20,7 @@ using Microsoft.Xna.Framework.Media;
 using System.IO;
 using windows_client.Languages;
 using System.Diagnostics;
+using System.Collections.Concurrent;
 
 namespace windows_client
 {
@@ -30,7 +31,7 @@ namespace windows_client
         bool WaitingForNonVoiceBlockingNetwork;
         bool WaitingForWiFi;
         IEnumerable<BackgroundTransferRequest> transferRequests;
-        private static Dictionary<string, ConvMessage> requestIdConvMsgMap = new Dictionary<string, ConvMessage>();
+        private static ConcurrentDictionary<string, ConvMessage> msgIdConvMsgMap = new ConcurrentDictionary<string, ConvMessage>();
 
         private static volatile FileTransfer instance = null;
         private static object syncRoot = new Object(); // this object is used to take lock while creating singleton
@@ -46,7 +47,7 @@ namespace windows_client
                         if (instance == null)
                         {
                             instance = new FileTransfer();
-                            instance.RemoveOldTransferRequests();
+                            instance.ProcessOldTransferRequests();
                         }
                     }
                 }
@@ -54,13 +55,15 @@ namespace windows_client
             }
         }
 
-        public void downloadFile(ConvMessage conMessage, string msisdn)
+        public void DownloadFile(ConvMessage conMessage, string msisdn)
         {
+            if (conMessage == null || string.IsNullOrEmpty(msisdn) || conMessage.FileAttachment == null)
+                return;
             Uri downloadUriSource = new Uri(Uri.EscapeUriString(HikeConstants.FILE_TRANSFER_BASE_URL + "/" + conMessage.FileAttachment.FileKey),
                 UriKind.RelativeOrAbsolute);
 
-            string relativeFilePath = "/" + msisdn + "/" + Convert.ToString(conMessage.MessageId);
-            string destinationPath = "shared/transfers" + "/" + Convert.ToString(conMessage.MessageId);
+            string relativeFilePath = msisdn + "/" + conMessage.MessageId;
+            string destinationPath = "shared/transfers" + "/" + conMessage.MessageId;
             Uri destinationUri = new Uri(destinationPath, UriKind.RelativeOrAbsolute);
 
             BackgroundTransferRequest transferRequest = new BackgroundTransferRequest(downloadUriSource);
@@ -71,11 +74,34 @@ namespace windows_client
             transferRequest.TransferStatusChanged += new EventHandler<BackgroundTransferEventArgs>(transfer_TransferStatusChanged);
             transferRequest.TransferProgressChanged += new EventHandler<BackgroundTransferEventArgs>(transfer_TransferProgressChanged);
             transferRequest.DownloadLocation = destinationUri;
+            bool addedToDownload = false;
             try
             {
+
                 transferRequest.TransferPreferences = TransferPreferences.AllowCellularAndBattery;
-                BackgroundTransferService.Add(transferRequest);
-                requestIdConvMsgMap.Add(transferRequest.RequestId, conMessage);
+                if (msgIdConvMsgMap.Count < 25)//max 25 downloads in queue
+                {
+                    if (!msgIdConvMsgMap.ContainsKey(transferRequest.Tag))
+                    {
+                        BackgroundTransferService.Add(transferRequest);
+                        msgIdConvMsgMap.TryAdd(transferRequest.Tag, conMessage);
+                        addedToDownload = true;
+                    }
+                    else
+                    {
+                        Debug.WriteLine("Already added to queue, TAG:" + transferRequest.Tag);
+                    }
+                }
+                else
+                {
+                    if (conMessage.UserTappedDownload)
+                    {
+                        Deployment.Current.Dispatcher.BeginInvoke(() =>
+                            {
+                                MessageBox.Show(AppResources.Download_MaxFiles_Txt);
+                            });
+                    }
+                }
             }
             catch (InvalidOperationException ex)
             {
@@ -87,6 +113,14 @@ namespace windows_client
                 Debug.WriteLine("FileTransfer :: downloadFile : downloadFile, Exception : " + e.StackTrace);
                 MessageBox.Show(AppResources.FileTransfer_ErrorMsgBoxText);
             }
+            finally
+            {
+                if (!addedToDownload)
+                {
+                    conMessage.SetAttachmentState(Attachment.AttachmentState.FAILED_OR_NOT_STARTED);
+                    MiscDBUtil.UpdateFileAttachmentState(msisdn, conMessage.MessageId.ToString(), Attachment.AttachmentState.FAILED_OR_NOT_STARTED);
+                }
+            }
         }
 
         private void ProcessTransfer(BackgroundTransferRequest transfer)
@@ -94,67 +128,74 @@ namespace windows_client
             switch (transfer.TransferStatus)
             {
                 case TransferStatus.Completed:
-                    if (transfer.StatusCode == 200 || transfer.StatusCode == 206)
+                    ConvMessage convMessage;
+                    msgIdConvMsgMap.TryGetValue(transfer.Tag, out convMessage);
+                    if (RemoveTransferRequest(transfer, transfer.Tag) && transfer.TransferError == null && (transfer.StatusCode == 200 || transfer.StatusCode == 206))
                     {
-                        // Remove the transfer request in order to make room in the 
-                        // queue for more transfers. Transfers are not automatically
-                        // removed by the system.
-                        ConvMessage convMessage;
-                        requestIdConvMsgMap.TryGetValue(transfer.RequestId, out convMessage);
-
                         if (convMessage != null)
                             convMessage.SetAttachmentState(Attachment.AttachmentState.COMPLETED);
 
-                        RemoveTransferRequest(transfer.RequestId);
-
-                        //RemoveTransferRequest(transfer.RequestId);
-                        // In this example, the downloaded file is moved into the root
-                        // Isolated Storage directory
                         if (transfer.UploadLocation == null)
                         {
-                            using (IsolatedStorageFile isoStore = IsolatedStorageFile.GetUserStoreForApplication())
+                            try
                             {
-                                string destinationPath = HikeConstants.FILES_BYTE_LOCATION + transfer.Tag;
-                                string destinationDirectory = destinationPath.Substring(0, destinationPath.LastIndexOf("/"));
-
-                                if (isoStore.FileExists(destinationPath))
-                                    isoStore.DeleteFile(destinationPath);
-
-                                if (!isoStore.DirectoryExists(destinationDirectory))
-                                    isoStore.CreateDirectory(destinationDirectory);
-
-                                isoStore.MoveFile(transfer.DownloadLocation.OriginalString, destinationPath);
-                                isoStore.DeleteFile(transfer.DownloadLocation.OriginalString);
-
-                                if (convMessage != null && convMessage.FileAttachment.ContentType.Contains(HikeConstants.IMAGE))
+                                string[] data = transfer.Tag.Split('/');
+                                if (data.Length == 2)
                                 {
-                                    IsolatedStorageFileStream myFileStream = isoStore.OpenFile(destinationPath, FileMode.Open, FileAccess.Read);
-                                    MediaLibrary library = new MediaLibrary();
-                                    myFileStream.Seek(0, 0);
-                                    library.SavePicture(convMessage.FileAttachment.FileName, myFileStream);
-                                }
+                                    string msisdn = data[0];
+                                    string messageId = data[1];
+                                    string destinationPath = HikeConstants.FILES_BYTE_LOCATION + "/" + transfer.Tag;
+                                    string destinationDirectory = destinationPath.Substring(0, destinationPath.LastIndexOf("/"));
 
-                                if (convMessage != null)
-                                {
-                                    var currentPage = ((App)Application.Current).RootFrame.Content as NewChatThread;
-                                    if (currentPage != null)
+
+                                    using (IsolatedStorageFile isoStore = IsolatedStorageFile.GetUserStoreForApplication())
                                     {
-                                        currentPage.displayAttachment(convMessage, true);
+                                        if (isoStore.FileExists(destinationPath))
+                                            isoStore.DeleteFile(destinationPath);
+
+                                        if (!isoStore.DirectoryExists(destinationDirectory))
+                                            isoStore.CreateDirectory(destinationDirectory);
+
+                                        isoStore.MoveFile(transfer.DownloadLocation.OriginalString, destinationPath);
+                                        if (isoStore.FileExists(transfer.DownloadLocation.OriginalString))
+                                            isoStore.DeleteFile(transfer.DownloadLocation.OriginalString);
+                                        if (convMessage != null && convMessage.FileAttachment.ContentType.Contains(HikeConstants.IMAGE))
+                                        {
+                                            IsolatedStorageFileStream myFileStream = isoStore.OpenFile(destinationPath, FileMode.Open, FileAccess.Read);
+                                            MediaLibrary library = new MediaLibrary();
+                                            myFileStream.Seek(0, 0);
+                                            library.SavePicture(convMessage.FileAttachment.FileName, myFileStream);
+                                        }
+                                    }
+
+                                    MiscDBUtil.UpdateFileAttachmentState(msisdn, messageId, Attachment.AttachmentState.COMPLETED);
+                                    if (convMessage != null)
+                                    {
+                                        var currentPage = ((App)Application.Current).RootFrame.Content as NewChatThread;
+                                        if (currentPage != null && convMessage.UserTappedDownload)
+                                        {
+                                            currentPage.displayAttachment(convMessage);
+                                        }
                                     }
                                 }
                             }
-                        }
-                        else
-                        {
+                            catch (Exception ex)
+                            {
+                                Debug.WriteLine("SOURCE:{0},\nEXCEPTION:{1},\nTAG:{2},\nFILEKEY:{3}", transfer.DownloadLocation.OriginalString, ex.Message, transfer.Tag, transfer.RequestUri.ToString());
+                            }
                         }
                     }
                     else
                     {
-                        ConvMessage convMessage;
-                        requestIdConvMsgMap.TryGetValue(transfer.RequestId, out convMessage);
-
                         if (convMessage != null)
                             convMessage.SetAttachmentState(Attachment.AttachmentState.FAILED_OR_NOT_STARTED);
+                        string[] data = transfer.Tag.Split('/');
+                        if (data.Length == 2)
+                        {
+                            string msisdn = data[0];
+                            string messageId = data[1];
+                            MiscDBUtil.UpdateFileAttachmentState(msisdn, messageId, Attachment.AttachmentState.FAILED_OR_NOT_STARTED);
+                        }
                         try
                         {
                             //case file doesn't exists on server
@@ -170,7 +211,6 @@ namespace windows_client
                                         });
                                 }
                             }
-                            RemoveTransferRequest(transfer.RequestId);
                             // This is where you can handle whatever error is indicated by the
                             // StatusCode and then remove the transfer from the queue. 
                             if (transfer.TransferError != null)
@@ -207,18 +247,18 @@ namespace windows_client
         void transfer_TransferStatusChanged(object sender, BackgroundTransferEventArgs e)
         {
             ProcessTransfer(e.Request);
-
         }
 
         void transfer_TransferProgressChanged(object sender, BackgroundTransferEventArgs e)
         {
             ConvMessage convMessage;
-            requestIdConvMsgMap.TryGetValue(e.Request.RequestId, out convMessage);
+            msgIdConvMsgMap.TryGetValue(e.Request.Tag, out convMessage);
             if (convMessage != null)
             {
                 if (convMessage.FileAttachment.FileState != Attachment.AttachmentState.CANCELED)
                 {
                     convMessage.ProgressBarValue = e.Request.BytesReceived * 100 / e.Request.TotalBytesToReceive;
+                    convMessage.ProgressText = string.Format("{0} of {1}", Utils.ConvertToStorageSizeString(e.Request.BytesReceived), Utils.ConvertToStorageSizeString(e.Request.TotalBytesToReceive));
                 }
                 else
                 {
@@ -235,7 +275,7 @@ namespace windows_client
             }
         }
 
-        private void RemoveOldTransferRequests()
+        public void ProcessOldTransferRequests()
         {
             if (transferRequests != null)
             {
@@ -249,6 +289,7 @@ namespace windows_client
 
             foreach (var transfer in transferRequests)
             {
+                msgIdConvMsgMap[transfer.Tag] = null;
                 transfer.TransferStatusChanged -= transfer_TransferStatusChanged;
                 transfer.TransferStatusChanged += new EventHandler<BackgroundTransferEventArgs>(transfer_TransferStatusChanged);
 
@@ -276,20 +317,64 @@ namespace windows_client
             //}
         }
 
-
-        private void RemoveTransferRequest(string transferID)
+        public void RemoveAllTransferRequests()
         {
-            requestIdConvMsgMap.Remove(transferID);
+            IEnumerable<BackgroundTransferRequest> transferRequests = BackgroundTransferService.Requests;
+
+            foreach (var transfer in transferRequests)
+            {
+                RemoveTransferRequest(transfer, transfer.Tag);
+            }
+        }
+
+        public void RemoveTransfersForMsisdn(string reqMsisdn)
+        {
+            IEnumerable<BackgroundTransferRequest> transferRequests = BackgroundTransferService.Requests;
+
+            foreach (var transfer in transferRequests)
+            {
+                string[] data = transfer.Tag.Split('/');
+                if (data.Length == 2)
+                {
+                    string msisdn = data[0];
+                    if (reqMsisdn == msisdn)
+                        RemoveTransferRequest(transfer, transfer.Tag);
+                }
+            }
+        }
+
+        private bool RemoveTransferRequest(BackgroundTransferRequest transferToRemove, string mapKey)
+        {
+            ConvMessage conv;
+            msgIdConvMsgMap.TryRemove(mapKey, out conv);
+            bool isSucess = false;
             // Use Find to retrieve the transfer request with the specified ID.
-            BackgroundTransferRequest transferToRemove = BackgroundTransferService.Find(transferID);
             try
             {
                 BackgroundTransferService.Remove(transferToRemove);
+                isSucess = true;
             }
             catch (Exception ex)
             {
-                Debug.WriteLine("FileTransfer :: RemoveTransferRequest : process transfer, Exception : " + ex.StackTrace);
+                Debug.WriteLine("FileTransfer :: RemoveTransferRequest : process transfer, Exception :{0}, StackTrace:{1} ", ex.Message, ex.StackTrace);
             }
+            return isSucess;
+        }
+
+        public void UpdateConvMap(ConvMessage convMessage, string msisdn)
+        {
+            string key = msisdn + "/" + convMessage.MessageId.ToString(); ;
+            if (msgIdConvMsgMap.ContainsKey(key))
+            {
+                msgIdConvMsgMap[key] = convMessage;
+            }
+        }
+
+        public ConvMessage GetDownloadingMessage(string msisdn, string msgId, out bool containsMsg)
+        {
+            ConvMessage downladMessage;
+            containsMsg = msgIdConvMsgMap.TryGetValue(msisdn + "/" + msgId, out downladMessage);
+            return downladMessage;
         }
 
     }
